+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/maatapi/model/contributions/ConcorContribEntry.java
@@ -15,6 +15,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ConcorContribEntry {
-    private int concorContributionId;
+    private Long concorContributionId;
     private String xmlContent;
 }
